Simplify API
… by consolidating more endpoints into the same module

This also adds support for `GET /v1/files/:id/content`
